Sync with origin for the latest version of code

+++ extmod/uzlib/tinflate.c
@@ -464,7 +464,7 @@ static int tinf_inflate_block_data(TINF_DATA *d, TINF_TREE *lt, TINF_TREE *dt)
             }
         } else {
             /* catch trying to point before the start of dest buffer */
-            if (offs > d->dest - d->destStart) {
+            if (offs > (unsigned int)(d->dest - d->destStart)) {
                 return TINF_DATA_ERROR;
             }
             d->lzOff = -offs;
